# apps/backend/static/openapi.json
{
  "openapi": "3.0.2",
  "paths": {
...
    "/api/v1/communities": {
      "post": {
        "summary": "Create a community",
        "tags": [
          "Community"
        ],
        "parameters": [],
        "operationId": "createCommunity",
        "requestBody": {
          "description": "Body",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "maxLength": 50
                  },
                  "slug": {
                    "type": "string",
                    "maxLength": 50
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 1000,
                    "nullable": true
                  },
                  "avatarKey": {
                    "type": "string",
                    "nullable": true
                  },
                  "bannerKey": {
                    "type": "string",
                    "nullable": true
                  }
                },
                "required": [
                  "name",
                  "slug"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "201",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "adminId": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string",
                      "nullable": true
                    },
                    "avatarKey": {
                      "type": "string",
                      "nullable": true
                    },
                    "bannerKey": {
                      "type": "string",
                      "nullable": true
                    },
                    "membersCount": {
                      "type": "number"
                    },
                    "postsCount": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "id",
                    "createdAt",
                    "updatedAt",
                    "adminId",
                    "name",
                    "slug",
                    "description",
                    "avatarKey",
                    "bannerKey",
                    "membersCount",
                    "postsCount"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "summary": "Search / browse communities",
        "tags": [
          "Community"
        ],
        "parameters": [
          {
            "name": "cursorSortValue",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursorCreatedAt",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "created_at",
                "members_count"
              ]
            }
          },
          {
            "name": "order",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "operationId": "getCommunities",
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "communityId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "slug": {
                            "type": "string"
                          },
                          "communityName": {
                            "type": "string"
                          },
                          "communityAvatarKey": {
                            "type": "string"
                          },
                          "membersCount": {
                            "type": "number"
                          },
                          "postsCount": {
                            "type": "number"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "communityId",
                          "slug",
                          "communityName",
                          "communityAvatarKey",
                          "membersCount",
                          "postsCount",
                          "createdAt"
                        ]
                      }
                    },
                    "cursorSortValue": {
                      "type": "string"
                    },
                    "cursorCreatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "hasMore": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "data",
                    "cursorSortValue",
                    "cursorCreatedAt",
                    "hasMore"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/communities/{idOrSlug}": {
      "get": {
        "summary": "Get a community by id or slug",
        "tags": [
          "Community"
        ],
        "parameters": [
          {
            "name": "idOrSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "operationId": "getCommunityByIdOrSlug",
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "adminId": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string",
                      "nullable": true
                    },
                    "avatarKey": {
                      "type": "string",
                      "nullable": true
                    },
                    "bannerKey": {
                      "type": "string",
                      "nullable": true
                    },
                    "membersCount": {
                      "type": "number"
                    },
                    "postsCount": {
                      "type": "number"
                    },
                    "viewerRole": {
                      "type": "string",
                      "enum": [
                        "member",
                        "moderator",
                        "admin"
                      ],
                      "nullable": true
                    }
                  },
                  "required": [
                    "id",
                    "createdAt",
                    "updatedAt",
                    "adminId",
                    "name",
                    "slug",
                    "description",
                    "avatarKey",
                    "bannerKey",
                    "membersCount",
                    "postsCount",
                    "viewerRole"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/communities/{id}": {
      "patch": {
        "summary": "Update community settings",
        "tags": [
          "Community"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "operationId": "updateCommunitySettings",
        "requestBody": {
          "description": "Body",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "maxLength": 50,
                    "nullable": true
                  },
                  "slug": {
                    "type": "string",
                    "maxLength": 50,
                    "nullable": true
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 1000,
                    "nullable": true
                  },
                  "avatarKey": {
                    "type": "string",
                    "nullable": true
                  },
                  "bannerKey": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "adminId": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string",
                      "nullable": true
                    },
                    "avatarKey": {
                      "type": "string",
                      "nullable": true
                    },
                    "bannerKey": {
                      "type": "string",
                      "nullable": true
                    },
                    "membersCount": {
                      "type": "number"
                    },
                    "postsCount": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "id",
                    "createdAt",
                    "updatedAt",
                    "adminId",
                    "name",
                    "slug",
                    "description",
                    "avatarKey",
                    "bannerKey",
                    "membersCount",
                    "postsCount"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "Delete a community",
        "tags": [
          "Community"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "operationId": "deleteCommunity",
        "responses": {
          "204": {
            "description": "204"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/communities/{id}/join": {
      "post": {
        "summary": "Join a community",
        "tags": [
          "Community"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "operationId": "joinCommunity",
        "responses": {
          "201": {
            "description": "201",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userId": {
                      "type": "string"
                    },
                    "communityId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "role": {
                      "type": "string",
                      "enum": [
                        "member",
                        "moderator",
                        "admin"
                      ]
                    },
                    "joinedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "userId",
                    "communityId",
                    "role",
                    "joinedAt",
                    "updatedAt"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/communities/{id}/leave": {
      "post": {
        "summary": "Leave a community",
        "tags": [
          "Community"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "operationId": "leaveCommunity",
        "responses": {
          "204": {
            "description": "204"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/communities/{id}/members/{userId}/role": {
      "patch": {
        "summary": "Change a community member's role",
        "tags": [
          "Community"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "operationId": "changeMemberRole",
        "requestBody": {
          "description": "Body",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "newRole": {
                    "type": "string",
                    "enum": [
                      "member",
                      "moderator",
                      "admin"
                    ]
                  }
                },
                "required": [
                  "newRole"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userId": {
                      "type": "string"
                    },
                    "communityId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "role": {
                      "type": "string",
                      "enum": [
                        "member",
                        "moderator",
                        "admin"
                      ]
                    },
                    "joinedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "userId",
                    "communityId",
                    "role",
                    "joinedAt",
                    "updatedAt"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/communities/{id}/bans": {
      "post": {
        "summary": "Ban a member from a community (moderator)",
        "tags": [
          "Community"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "operationId": "banMember",
        "requestBody": {
          "description": "Body",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userIdToBan": {
                    "type": "string"
                  }
                },
                "required": [
                  "userIdToBan"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "201",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "communityId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "userId": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "communityId",
                    "userId",
                    "createdAt"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "version": "1.0.0",
    "title": "Nexus REST API - Documentation",
    "description": "Nexus REST API - Documentation"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "x-service-token": {
        "type": "apiKey",
        "name": "x-service-token",
        "in": "header"
      }
    }
  }
}